proof_of_concept/logger_agent/pyproject.toml
[project]
name = "logger-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-arango>=8.2.1",
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "litellm>=1.0.0",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "openai>=1.0.0",
    "psutil>=7.0.0",
    "python-arango>=8.2.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "setuptools>=80.9.0",
    "tenacity>=9.1.2",
    "urllib3<2",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
    "websockets>=15.0.1",
    "httpx>=0.28.1",
]
